+++ migrations/20150422195559_credenciado.js
@@ -2,7 +2,7 @@
 
 exports.up = function(knex, Promise) {
   return knex.schema.createTable('credenciado', function(table) {
-    table.bigInteger('pessoa_id')
+    table.integer('pessoa_id')
       .primary()
       .references('id_pessoa').inTable('pessoa');
     table.boolean('contrato_de_servico_valido');
