+++ migrations/20150422195616_usuario_revendedor.js
@@ -2,7 +2,7 @@
 
 exports.up = function(knex, Promise) {
   return knex.schema.createTable('usuario_revendedor', function(table) {
-    table.bigInteger('pessoa_fisica_pessoa_id')
+    table.integer('pessoa_fisica_pessoa_id')
       .primary()
       .references('pessoa_id').inTable('pessoa_fisica');
     table.string('login').unique().notNullable();
@@ -10,7 +10,7 @@ exports.up = function(knex, Promise) {
     table.boolean('primeiro_acesso').notNullable();
     table.string('autorizacao').notNullable();
     table.boolean('ativo').notNullable();
-    table.bigInteger('revendedor_id')
+    table.integer('revendedor_id')
       .references('pessoa_id').inTable('revendedor');
   });
 };
