+++ migrations/20150511122507_recuperacao_senha.js
@@ -4,7 +4,7 @@ exports.up = function(knex, Promise) {
   return knex.schema.createTable('recuperacao_senha', function(table) {
     table.uuid('id_recuperacao_senha').notNullable();
     table.timestamp('data_expiracao').notNullable();
-    table.bigInteger('pessoa_id').notNullable()
+    table.integer('pessoa_id').notNullable()
       .references('id_pessoa').inTable('pessoa');
   });
 };
